Add MssqlTarget to backup operator
Added a new MssqlTarget to klusoga-backup-operator and updated relevant files. The MSSQLTarget is a new custom resource for a backup solution and allows the controller to handle backup processes for Microsoft SQL targets. New configurations were added to the role settings to permit necessary CRUD operations. Necessary boilerplate code and examples were also generated for the MssqlTarget.
This addition expands the capability of this backup solution to cover Microsoft SQL servers.