Enhance MSSQLTarget API with additional fields
The MSSQLTarget YAML, API and Dockerfile were updated to include new fields: 'AppLabel', 'CredentialsRef', 'DestinationRef' and 'Databases'. These changes enhance the specification of the MSSQLTarget for users, allowing them to define necessary parameters for the MSSQLTarget including application label, reference to credentials, destination and array of databases. This greatly improves the flexibility and customization of the MSSQLTarget API. Changes have been mirrored in the Dockerfile for consistency.