"Extend destination and controller functionality
Added support to properly create and reconcile Destination resource type. This is achieved by enhancing the DestinationReconciler and MssqlTargetReconciler in the controllers. Now it can handle situation where destination not found and create DestinationConfigmap. Moreover, alterations in the SQL credentials fetching logic and scenarios to cope with Secret and Destination resources are implemented. These changes ensure better error handling and cleaner resource management."