"Update README with klusoga Backup Operator usage & MsSQLTarget details
README.md has been updated to reflect a clear understanding of Klusoga Backup Operator and the usage of MssqlTarget variable. The sweeping changes were necessary to improve user clarity and provide a quickstart guide for the project. An explanation of MssqlTarget has been included with a detailed tabulation of parameters enabling efficient backups of Microsoft SQL backups."

[ci-skip]